requirements.txt
Django==1.11.13
django-webpack-loader
django-sslify
mysqlclient==1.3.12
bcrypt
django-mysql==2.2.2
python-docx
comtypes
